refactor(agent_meta): extract pane + skills helpers to _meta/
Pure move: _SUBAGENT_MARKER_RE, parse_subagent_count_from_pane_text,
_subagent_count_from_pane, _capture_pane, _classify_pane_state now
live in _state/_meta/pane.py; _parse_skills in _state/_meta/skills.py.
agent_meta re-exports each so existing am._foo access keeps working.

Integer literals updated to PEP-515 separators (10_000 / 2_000) in the
moved code only — pure formatting improvement, no behavior change.

412 _state/ tests pass (one flake in test_port_allocator unrelated to
this change passes on re-run).
